fixed typo in ST_USyd_Aus files
